requirements.txt
asgiref==3.5.2
attrs==22.1.0
coverage==6.5.0
Django==4.1.4
django-autoslug==1.9.8
django-model-utils==4.3.1
factory-boy==3.2.1
Faker==15.3.4
iniconfig==1.1.1
packaging==22.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
sqlparse==0.4.3
